Function pallet_conviction_voting::pallet::dispatchables::remove_other_vote

pub fn remove_other_vote<T: Config<I>, I: 'static>(
    target: <<T as Config>::Lookup as StaticLookup>::Source,
    class: <<T as Config<I>>::Polls as Polling<TallyOf<T, I>>>::Class,
    index: <<T as Config<I>>::Polls as Polling<TallyOf<T, I>>>::Index
)

Remove a vote for a poll.

If the target is equal to the signer, then this function is exactly equivalent to remove_vote. If not equal to the signer, then the vote must have expired, either because the poll was cancelled, because the voter lost the poll or because the conviction period is over.

The dispatch origin of this call must be Signed.

target: The account of the vote to be removed; this account must have voted for poll index.

index: The index of poll of the vote to be removed.

class: The class of the poll.

Weight: O(R + log R) where R is the number of polls that target has voted on. Weight is calculated for the maximum number of vote.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::remove_other_vote.