Function pallet_conviction_voting::pallet::dispatchables::unlock

pub fn unlock<T: Config<I>, I: 'static>(
    class: <<T as Config<I>>::Polls as Polling<TallyOf<T, I>>>::Class,
    target: <<T as Config>::Lookup as StaticLookup>::Source
)

Remove the lock caused by prior voting/delegating which has expired within a particular class.

The dispatch origin of this call must be Signed.

class: The class of polls to unlock.

target: The account to remove the lock on.

Weight: O(R) with R number of vote of target.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the unlock variant of the pallet Call enum. You should not attempt to call this function directly.