Function pallet_conviction_voting::pallet::dispatchables::vote

pub fn vote<T: Config<I>, I: 'static>(
    poll_index: <<T as Config<I>>::Polls as Polling<TallyOf<T, I>>>::Index,
    vote: AccountVote<<<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance>,
)

Vote in a poll. If vote.is_aye(), the vote is to enact the proposal; otherwise it is a vote to keep the status quo.

The dispatch origin of this call must be Signed.

poll_index: The index of the poll to vote for.

vote: The vote configuration.

Weight: O(R) where R is the number of polls the voter has voted on.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::vote.