Type Definition pallet_conviction_voting::pallet::VotingFor

pub type VotingFor<T: Config<I>, I: 'static = ()> = StorageDoubleMap<_GeneratedPrefixForStorageVotingFor<T, I>, Twox64Concat, T::AccountId, Twox64Concat, <<T as Config<I>>::Polls as Polling<TallyOf<T, I>>>::Class, Voting<<<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance, <T as Config>::AccountId, <T as Config>::BlockNumber, <<T as Config<I>>::Polls as Polling<TallyOf<T, I>>>::Index, <T as Config<I>>::MaxVotes>, ValueQuery>;

All voting for a particular voter in a particular voting class. We store the balance for the number of votes that we have recorded.