pallet_conviction_voting

Type Alias TallyOf

pub type TallyOf<T, I = ()> = Tally<BalanceOf<T, I>, <T as Config<I>>::MaxTurnout>;

Aliased Type

pub struct TallyOf<T, I = ()> {
    pub ayes: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    pub nays: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    pub support: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    /* private fields */
}

Fields
ayes: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance

The number of aye votes, expressed in terms of post-conviction lock-vote.

nays: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance

The number of nay votes, expressed in terms of post-conviction lock-vote.

support: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance

The basic number of aye votes, expressed pre-conviction.