Crate pallet_conviction_voting

Voting Pallet

Config

Call

Overview

Pallet for managing actual voting in polls.

Re-exports

pub use self::weights::WeightInfo;

pub use self::pallet::*;

Modules

benchmarking

ConvictionVoting pallet benchmarking.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_conviction_voting

Structs

Casting

Information concerning the direct vote-casting of some voting power.

Delegating

Information concerning the delegation of some voting power.

Delegations

Amount of votes and capital placed in delegation for an account.

Tally

Info regarding an ongoing referendum.

Vote

A number of lock periods, plus a vote, one way or the other.

Enums

AccountVote

A vote for a referendum of a particular account.

Conviction

A value denoting the strength of conviction of a vote.

UnvoteScope

Whether an unvote operation is able to make actions that are not strictly always in the interest of an account.

Voting

An indicator for what an account is doing; it can either be delegating or voting.

Type Aliases

BalanceOf

BlockNumberFor

ClassOf

IndexOf

PollIndexOf

TallyOf

VotesOf

VotingOf