pallet_conviction_voting::pallet

Trait Config

pub trait Config<I: 'static = ()>:
    Config
    + Sized
    + Config<RuntimeEvent: From<Event<Self, I>>> {
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type WeightInfo: WeightInfo;
    type Currency: ReservableCurrency<Self::AccountId> + LockableCurrency<Self::AccountId, Moment = BlockNumberFor<Self, I>> + Inspect<Self::AccountId>;
    type Polls: Polling<TallyOf<Self, I>, Votes = BalanceOf<Self, I>, Moment = BlockNumberFor<Self, I>>;
    type MaxTurnout: Get<BalanceOf<Self, I>>;
    type MaxVotes: Get<u32>;
    type VoteLockingPeriod: Get<BlockNumberFor<Self, I>>;
    type BlockNumberProvider: BlockNumberProvider;
    type VotingHooks: VotingHooks<Self::AccountId, PollIndexOf<Self, I>, BalanceOf<Self, I>>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type Currency: ReservableCurrency<Self::AccountId> + LockableCurrency<Self::AccountId, Moment = BlockNumberFor<Self, I>> + Inspect<Self::AccountId>

Currency type with which voting happens.

type Polls: Polling<TallyOf<Self, I>, Votes = BalanceOf<Self, I>, Moment = BlockNumberFor<Self, I>>

The implementation of the logic which conducts polls.

type MaxTurnout: Get<BalanceOf<Self, I>>

The maximum amount of tokens which may be used for voting. May just be Currency::total_issuance, but you might want to reduce this in order to account for funds in the system which are unable to vote (e.g. parachain auction deposits).

type MaxVotes: Get<u32>

The maximum number of concurrent votes an account may have.

Also used to compute weight, an overly large value can lead to extrinsics with large weight estimation: see delegate for instance.

type VoteLockingPeriod: Get<BlockNumberFor<Self, I>>

The minimum period of vote locking.

It should be no shorter than enactment period to ensure that in the case of an approval, those successful voters are locked into the consequences that their votes entail.

type BlockNumberProvider: BlockNumberProvider

Provider for the block number. Normally this is the frame_system pallet.

type VotingHooks: VotingHooks<Self::AccountId, PollIndexOf<Self, I>, BalanceOf<Self, I>>

Hooks are called when a new vote is registered or an existing vote is removed.

The trait does not expose weight information. The weight of each hook is assumed to be benchmarked as part of the function that calls it. Hooks should never recursively call into functions that called, directly or indirectly, the function that called them. This could lead to infinite recursion and stack overflow. Note that this also means to not call into other generic functionality like batch or similar. Also, anything that a hook did will be subject to the transactional semantics of the calling function. This means that if the calling function fails, the hook will be rolled back without further notice.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.