Type Alias pallet_conviction_voting::TallyOf

pub type TallyOf<T, I = ()> = Tally<<<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance, <T as Config<I>>::MaxTurnout>;

Aliased Type

struct TallyOf<T, I = ()> {
    pub ayes: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    pub nays: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    pub support: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    /* private fields */
}

Fields
ayes: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance

The number of aye votes, expressed in terms of post-conviction lock-vote.

nays: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance

The number of nay votes, expressed in terms of post-conviction lock-vote.

support: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance

The basic number of aye votes, expressed pre-conviction.

Implementations

impl<Votes: Clone + Default + PartialEq + Eq + Debug + Copy + AtLeast32BitUnsigned + TypeInfo + Codec, Total: Get<Votes>> Tally<Votes, Total>

pub fn from_vote(vote: Vote, balance: Votes) -> Self

Create a new tally.

pub fn from_parts( ayes_with_conviction: Votes, nays_with_conviction: Votes, ayes: Votes ) -> Self

pub fn add(&mut self, vote: AccountVote<Votes>) -> Option<()>

Add an account’s vote into the tally.

pub fn remove(&mut self, vote: AccountVote<Votes>) -> Option<()>

Remove an account’s vote from the tally.

pub fn increase(&mut self, approve: bool, delegations: Delegations<Votes>)

Increment some amount of votes.

pub fn reduce(&mut self, approve: bool, delegations: Delegations<Votes>)

Decrement some amount of votes.